requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.2
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
drf-spectacular-sidecar==2023.3.1
factory-boy==3.2.1
Faker==17.6.0
filelock==3.9.0
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
identify==2.5.18
inflection==0.5.1
isort==5.12.0
jsonschema==4.17.3
kombu==5.2.4
mccabe==0.7.0
nodeenv==1.7.0
Pillow==9.4.0
platformdirs==3.1.0
pre-commit==3.1.1
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
vine==5.0.0
virtualenv==20.20.0
wcwidth==0.2.6
